Delete the forecast duplicate code
